Refactor(step4) Race 객체 내부에 Car객체 리스트를 포함하도록 수정
