+++ src/main/kotlin/step4/CarRacing.kt
@@ -1,5 +1,6 @@
 package step4
 
+import step4.model.Race
 import step4.view.InputView
 import step4.view.ResultView
 
@@ -7,12 +8,14 @@ fun main() {
     val nameListOfCars = InputView.getNameListOfCars()
     val numberOfTrials = InputView.getNumberOfTrials()
 
-    var cars = Race.initiate(nameListOfCars)
+    val races = Race(nameListOfCars)
 
     repeat(numberOfTrials) {
-        cars = Race.makeTurn(cars)
+        races.makeTurn()
     }
 
-    ResultView.output(numberOfTrials, cars)
-    ResultView.noticeWinner(cars)
+    val carsAfterRace = races.getCars()
+
+    ResultView.output(numberOfTrials, carsAfterRace)
+    ResultView.noticeWinner(carsAfterRace)
 }
