+++ src/main/kotlin/step4/model/Car.kt
@@ -1,4 +1,4 @@
-package step4
+package step4.model
 
 data class Car(val name: String, var raceResult: Int, val raceHistory: MutableList<Int>) {
     fun saveResult(resultOfTurn: Int) {
